se hicieron ultimos detalles

+++ QuadTreeADT.h
@@ -9,5 +9,5 @@ public:
     virtual void insert(Point p, int data)=0;   //Inserta un nuevo punto p en el QuadTree, asociando a dicho punto la información disponible en data (ej. la población de una ciudad con coordenadas p).
     virtual vector<Node*> list()=0;     //Retorna un contenedor con todos los puntos en el QuadTree. Por cada punto retorna coordenadas y su valor data(usar preorder o otros para mostrar)
     virtual int countRegion(Point p, int d)=0;  //Retorna la cantidad de puntos en una región del plano, tomando como centro el punto p y una distancia d.
-    virtual int AggregateRegion(Point p, int d)=0;  //Retorna la población estimada dentro de una región del plano, tomando como centro el punto p y una distancia d.
+    virtual long long AggregateRegion(Point p, int d)=0;  //Retorna la población estimada dentro de una región del plano, tomando como centro el punto p y una distancia d.
 };
