+++ quadtree.h
@@ -55,7 +55,7 @@ class Quad: public Quadtree{
     void _list(Quad* quad, vector<Node*>& lista);
     int countRegion(Point p, int d);
     void _countRegion(Quad* quad, int minX, int maxX, int minY, int maxY, int& count);
-    int AggregateRegion(Point p, int d);
-    void _AggregateRegion(Quad* quad, int minX, int maxX, int minY, int maxY, int& count);
+    long long AggregateRegion(Point p, int d);
+    void _AggregateRegion(Quad* quad, int minX, int maxX, int minY, int maxY, long long& count);
 
 };
